refactor(coordinateur): CoordinatorEvaluations.vue sous 300 lignes (H2)
Éclatement complet (pattern G1) :
- useCoordinatorEvaluations.js : état + chargement global + filtres/stats + nav
- CoordinatorEvalStats.vue : 4 cartes de stats (CSS verbatim)
- CoordinatorEvalFilters.vue : filtres enseignant/classe/matière/statut (v-model + emits)
- CoordinatorEvalCard.vue : carte d'évaluation + emits view-results/view-details (CSS verbatim)
- CoordinatorEvaluations.vue : orchestrateur 801→188 lignes
Tests : CoordinatorEvaluations.test.js (existant, 2) + composable (5) + sections (5) verts.
Build OK. Parité : mêmes endpoints (/evaluations, /proxy/classes, /proxy/matieres) et routes.
